PYTHON/TREASURE_HUNTER_1/requirements.txt
colorama==0.4.6
keyboard==0.13.5
pygame-ce==2.5.2
PyTMX==3.32
setuptools==75.6.0
termcolor==2.5.0
wheel==0.45.1
